Trekk ut tilpasning av utdypende vilkår for automatisk kjøring av Finnmark og Svalbard til egen fil
